requirements.txt
aiohttp==3.6.2
arrow==0.16.0
async-timeout==3.0.1
attrs==20.2.0
base58==2.0.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==6.6
docker-py==1.8.0
ecdsa==0.16.0
flake8==3.8.3
funcsigs==1.0.2
future==0.18.2
idna==2.10
iniconfig==1.0.1
jsonrpcclient==2.0.1
jsonrpcserver==3.1.1
jsonschema==3.2.0
mccabe==0.6.1
mnemonic==0.19
more-itertools==8.5.0
multidict==4.7.6
packaging==20.4
path==15.0.0
path.py==12.5.0
pexpect==4.8.0
pluggy==0.13.1
protobuf==3.13.0
ptyprocess==0.6.0
py==1.9.0
pyaes==1.6.1
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.8
pyflakes==2.2.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==6.0.1
python-binance-chain==0.1.20
python-dateutil==2.8.1
pywallet==0.1.0
PyYAML==5.3.1
requests==2.24.0
secp256k1==0.13.2
sha256==0.1
six==1.15.0
tabulate==0.8.7
toml==0.10.1
two1==3.10.9
ujson==3.1.0
urllib3==1.25.10
websocket-client==0.57.0
websockets==8.1
yarl==1.5.1
